requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg==3.1.19
psycopg2-binary==2.9.9
PyJWT==2.10.1
pytest==8.3.4
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
